fb472 — THE SPECTRAL CUT MOVES ONTO THE BACK PANEL'S LOW AND HIGH KNOBS
Max: "not a new spectral, not two new spectrum modes, especially not two on one button... wire in
the low and the high cut in the back channel where it says low and high, because I thought that's
what it was for."

He is right, and the mismatch was mine. fb467 put a partial WINDOW on those two knobs — which
partials the morph was allowed to touch — an abstraction nobody asked for, sitting under labels
that promise a filter. fb471 then added the actual cut as two SPECTRAL MODES, which crowded the
mode list and forced a choice between a cut and a morph.

WHAT CHANGED
 · the two `SpectralMode` entries are GONE; the list is back to nine.
 · SYN_OSC_x_SPECTRAL_LO / _HI keep their ids and their modulation destinations (SpecLoA..D = 1846,
   SpecHiA..D = 1850) and become a fourth-order Butterworth LOW CUT and HIGH CUT in HARMONIC NUMBER
   — the corner rides the note, where Serum 2's spectral markers are in hertz and do not track.
 · the params are plain 0..1 with NO skew, because the corner is already exponential in the knob
   (0.25*2^11t and 0.5*2^13t): a linear knob gives a logarithmic sweep. Low at 0 and High at 1 SKIP
   the filter outright — an exact identity, which parking a corner at harmonic 1 would not be,
   since a fourth-order response is still -3 dB at its own corner. Those ends read "off".

TWO THINGS THE MOVE REQUIRED, and either would have made it do nothing:
 · a cut must apply with mode = None. `apply()` returned `base` on None, which is the whole reason
   a cut belongs on knobs rather than in the mode list.
 · `rebuildMorphIfNeeded` published nullptr unless mode > 0 AND amount > 0, so a cut on its own
   never baked. It bakes for a cut alone now, and the corner is quantised to 1/512 so a modulated
   cut cannot churn a 2.3 ms bake at 60 Hz.

MEASURED: corner -3.01 dB, two octaves below -48.2 dB, clean above to -0.000, mirrored for the high
cut, zero ratio and zero phase drift over 40 partials, and a cut STACKS with a morph mode rather
than replacing one (low-partial sum 1.0000 -> 0.0002 with Harmonic Stretch still engaged). On the
installed AU, with NO morph mode: low cut -8.9 dB from clean, high cut -3.0, the two -0.8 dB apart
from each other, both returned to off -69.2 dB against a -72.2 dB repeatability floor, and an LFO
on the high cut moves it by -15.6 dB.

The fb467 window's gates (W1-W5) are RETIRED IN PLACE with the reason recorded rather than deleted
quietly, and spec_cert N5 — which held the JS readout to the C++ mapping — is re-pointed at the
cut's corner constants so the two files still cannot drift.

spec_cert 16/16 · eq_cut_cert 9/9 · au_spec 14/14 · eq_ui 13/13 · fx4_ui 131/131 ·
blur_twin_cert 8/8 · wtfft_cert 6/6 · fxmod_cert 37/37 · fxmod_move 58/58 · fxmod_menu 19/19 ·
auval 0 · pluginval both.

+++ plugins/TerrainInstrument/Source/PluginProcessor.h
@@ -618,7 +618,7 @@ public:
     // audio thread publishes each block (fb252/fb76: base + LFO/env, quantised to 1/128 only when
     // routed), which is the same number rebuildMorphIfNeeded builds from — so the picture the
     // waterfall re-bakes is the morph the ear is hearing, not the knob.
-    void spectralDisplay (int o, float& amtOut, int& typeOut, int& loOut, int& hiOut) const noexcept;
+    void spectralDisplay (int o, float& amtOut, int& typeOut, float& loOut, float& hiOut) const noexcept;
     juce::uint32 modDragSeq_ = 0;
     static bool physicalLeftButtonDown();               // fb151 — window-server button truth (see PluginProcessor.cpp)
     void adoptCardWindow (const juce::String& id, std::unique_ptr<juce::Component> w);
@@ -1373,8 +1373,8 @@ private:
         // fb467 — the partial WINDOW is part of the built table's identity. Without these two the
         // skip gate below matches on preset/mode/amount alone and Lo/Hi move NOTHING until some
         // other knob happens to force a rebuild — the failure that builds clean and looks wired.
-        int   builtLo     = -1;
-        int   builtHi     = -1;
+        float builtLo     = -1.0f;
+        float builtHi     = -1.0f;
         const tw::Wavetable* builtImportPtr = nullptr;   // fb253 — morph SOURCE was this import (nullptr = a factory preset)
         int   builtImportEpoch = -1;                     // fb253 — the import's buildEpoch when morphed (re-import → re-morph)
     };
